@@ -3,10 +3,11 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "react": "^16.1.0",
+    "material-ui": "^0.20.0",
+    "react": "^16.2.0",
     "react-async-script": "^0.9.1",
+    "react-copy-to-clipboard": "^5.0.1",
     "react-dom": "^16.1.0",
-    "react-google-recaptcha": "^0.8.1",
     "react-router-dom": "^4.2.2",
     "react-scripts": "1.0.17",
     "whatwg-fetch": "^2.0.3"
